remove HasReqsHints import cycle
